# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API documentation for FAS Management System.",
        "title": "FAS Management System API",
        "contact": {},
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "http://localhost:8080",
    "basePath": "/api",
    "paths": {
        "/applicants": {
            "get": {
                "description": "Retrieves and returns a list of all registered applicants.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "List All Applicants",
                "responses": {
                    "200": {
                        "description": "Successfully retrieved list of applicants.",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/http.ApplicantResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Handles the creation of a new applicant by accepting necessary data and storing it in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "Create a new Applicant",
                "parameters": [
                    {
                        "description": "Payload for creating a new applicant",
                        "name": "CreateApplicantRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.CreateApplicantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created applicant.",
                        "schema": {
                            "$ref": "#/definitions/http.ApplicantResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/applicants/{id}": {
            "get": {
                "description": "Retrieves the details of a single applicant using their unique identifier.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "Retrieve Applicant by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Applicant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully retrieved applicant.",
                        "schema": {
                            "$ref": "#/definitions/http.ApplicantResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Applicant Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the specified applicant's details based on the provided payload.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "Update an Applicant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Applicant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload for updating an applicant",
                        "name": "UpdateApplicantRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UpdateApplicantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated applicant.",
                        "schema": {
                            "$ref": "#/definitions/http.ApplicantResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Applicant Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes the applicant with the specified ID from the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Applicants"
                ],
                "summary": "Delete an Applicant",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Applicant ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted applicant.",
                        "schema": {
                            "$ref": "#/definitions/http.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Applicant Not Found",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/http.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.EmploymentStatus": {
            "type": "string",
            "enum": [
                "employed",
                "unemployed"
            ],
            "x-enum-varnames": [
                "EmploymentStatusEmployed",
                "EmploymentStatusUnemployed"
            ]
        },
        "domain.MaritalStatus": {
            "type": "string",
            "enum": [
                "single",
                "married",
                "widowed",
                "divorce"
            ],
            "x-enum-varnames": [
                "MaritalStatusSingle",
                "MaritalStatusMarried",
                "MaritalStatusWidowed",
                "MaritalStatusDivorce"
            ]
        },
        "domain.Sex": {
            "type": "string",
            "enum": [
                "male",
                "female"
            ],
            "x-enum-varnames": [
                "SexMale",
                "SexFemale"
            ]
        },
        "http.ApplicantResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                },
                "date_of_birth": {
                    "type": "string",
                    "example": "2000-01-01"
                },
                "employment_status": {
                    "type": "string",
                    "example": "employed"
                },
                "id": {
                    "type": "string",
                    "example": "00000000-0000-0000-0000-000000000000"
                },
                "marital_status": {
                    "type": "string",
                    "example": "married"
                },
                "name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "sex": {
                    "type": "string",
                    "example": "male"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2021-01-01T00:00:00Z"
                }
            }
        },
...
        "http.CreateApplicantRequest": {
            "type": "object",
            "required": [
                "date_of_birth",
                "employment_status",
                "marital_status",
                "name",
                "sex"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "employment_status": {
                    "$ref": "#/definitions/domain.EmploymentStatus"
                },
                "marital_status": {
                    "$ref": "#/definitions/domain.MaritalStatus"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "$ref": "#/definitions/domain.Sex"
                }
            }
        },
...
        "http.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "field": "error description"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Error message"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "http.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "http.UpdateApplicantRequest": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "date_of_birth": {
                    "type": "string"
                },
                "employment_status": {
                    "$ref": "#/definitions/domain.EmploymentStatus"
                },
                "id": {
                    "type": "string"
                },
                "marital_status": {
                    "$ref": "#/definitions/domain.MaritalStatus"
                },
                "name": {
                    "type": "string"
                },
                "sex": {
                    "$ref": "#/definitions/domain.Sex"
                }
            }
        },
...
    }
}